Changes the result metric to not embed Order
By embedding order, we made it impossible to use transformers on
metadata.Order more or less, which created some really confusing bugs
since some transformers worked, but not others.

In addition to this change, some minor style changes are included, and
bumping Skogul to v0.25.1 to get the new dictsplit-behavior, making it
trivial to split things up.