Make smierte global, and move orders around
gosmi works globally, and it seems pointless to then insist on passing a
local smierte.Config around to do lookups.

The reason I did this along with moving orders around is that I
discovered a problem with walks being completely broken, then just
somewhat broken when I moved the orders around and went to just test
one. Sorry about the mess.

This ALSO ensures that walks: 1. don't crash outright, 2. get the
benefit of lookups.

I see now, though, that there's no point doing this mapback-stuff, and I
should just move the type-lookup to the callback and out of Get and
BulkWalk.